--- Python/pyth.py
@@ -1,8 +1,0 @@
-from . smth import newClass
-
-class newClass (
-    newClass = 3
-)
-
-def greet(name, msg):
-    print("Kate" + ',' +"how are you?")
