--- Python/smth.py
@@ -1,34 +1,0 @@
-first_name = "Aneta"
-second_name = " Vankova"
-
-print(first_name + second_name )
-
-
-active_users = 5
-number_of_applications = 6 + 1 
-
-print(active_users )
-print(number_of_applications)
-
-
-status = "Watching Netflix"
-status = "Relaxing at the beach"
-print(status)
-
-
-default_option = "upload"
-new_status = "download"
-
-new_status = default_optionm
-print(new_status)
-
-powered_on = True 
-
-correct = True
-not_correct = False
-print(correct)
-print(not_correct)
-
-print(not True)
-print(not False)
-
